requirements.txt
fastapi
Pillow
pytesseract
xmltodict
docx2txt
PyPDF2
uvicorn
python-multipart
